Create ONE AST from parsing multiple parseCompilationUnits.
We need this in order to be able to get bindings. Otherwise, some IMethodBinding or IVariableBinding have a null key.
This may make the framework less scalable. But is kind of needed for whole program analyses.

Otherwise, a whole program analysis will have compensate for this, which may be equally bad.